FIX: Make memos list not to be rendered if memos are empty
